break apart main.go into logical file separations
