+++ internal/middleware/middleware.go
@@ -1,0 +1,10 @@
+package middleware
+
+import "net/http"
+
+func Chain(handler http.Handler, middlewares ...func(http.Handler) http.Handler) http.Handler {
+	for i := len(middlewares) - 1; i >= 0; i-- {
+		handler = middlewares[i](handler)
+	}
+	return handler
+}
