+++ internal/middleware/ratelimit.go
@@ -1,0 +1,58 @@
+package middleware
+
+import (
+	"net/http"
+	"strings"
+	"sync"
+	"time"
+
+	"golang.org/x/time/rate"
+)
+
+type RateLimiter struct {
+	visitors map[string]*rate.Limiter
+	mu       sync.RWMutex
+}
+
+func NewRateLimiter() *RateLimiter {
+	return &RateLimiter{
+		visitors: make(map[string]*rate.Limiter),
+	}
+}
+
+func (rl *RateLimiter) getLimiter(ip string) *rate.Limiter {
+	rl.mu.RLock()
+	limiter, exists := rl.visitors[ip]
+	rl.mu.RUnlock()
+
+	if !exists {
+		rl.mu.Lock()
+		limiter = rate.NewLimiter(rate.Every(time.Second), 10)
+		rl.visitors[ip] = limiter
+		rl.mu.Unlock()
+	}
+
+	return limiter
+}
+
+func (rl *RateLimiter) Allow(ip string) bool {
+	return rl.getLimiter(ip).Allow()
+}
+
+var rateLimiter = NewRateLimiter()
+
+func RateLimitMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		ip := r.RemoteAddr
+		if forwarded := r.Header.Get("X-Forwarded-For"); forwarded != "" {
+			ip = strings.Split(forwarded, ",")[0]
+		}
+
+		if !rateLimiter.Allow(strings.TrimSpace(ip)) {
+			http.Error(w, "Rate limit exceeded", http.StatusTooManyRequests)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	})
+}
