+++ internal/models/extension.go
@@ -1,0 +1,8 @@
+package models
+
+type Extension struct {
+	ID          string `json:"id"`
+	Name        string `json:"name"`
+	Description string `json:"description"`
+	Enabled     bool   `json:"enabled"`
+}
